backend-pipeline/requirements.txt
django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
celery>=5.3.0
redis>=4.5.0
flower>=2.0.0
psycopg2-binary>=2.9.0
requests>=2.30.0
django-celery-beat>=2.5.0
python-dotenv>=1.0.0
feedparser>=6.0.0

# New dependencies for enhanced analytics
textblob>=0.17.1
spacy>=3.7.2
transformers>=4.36.2
torch>=2.1.2
openai>=1.3.7
numpy>=1.24.3
pandas>=2.1.4
scikit-learn>=1.3.2
GeoDjango>=4.2.0
geopy>=2.4.1
folium>=0.14.0
plotly>=5.18.0
dash>=2.14.2
dash-bootstrap-components>=1.5.0 